Product repository was created. Id type of Product entity was changed from BigInteger to Long.

+++ src/test/java/com/sergosoft/intergalacticmarketplace/entity/Product.java
@@ -10,15 +10,13 @@ import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Positive;
 
-import java.math.BigInteger;
-
 @Entity
 @Data
 public class Product {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private BigInteger id;
+    private Long id;
 
     @NotBlank
     private String title;
